+++ lib/src/column.dart
@@ -6,6 +6,7 @@ import 'package:davi/src/cell_semantics_builder.dart';
 import 'package:davi/src/column_id.dart';
 import 'package:davi/src/pin_status.dart';
 import 'package:davi/src/sort.dart';
+import 'package:davi/src/summary_builder.dart';
 import 'package:davi/src/value_mapper.dart';
 import 'package:flutter/semantics.dart';
 import 'package:flutter/widgets.dart';
@@ -35,11 +36,12 @@ class DaviColumn<DATA> extends ChangeNotifier {
       this.leading,
       DaviDataComparator<DATA>? dataComparator,
       this.pinStatus = PinStatus.none,
-      DaviIntValueMapper<DATA>? intValue,
-      DaviDoubleValueMapper<DATA>? doubleValue,
-      DaviStringValueMapper<DATA>? stringValue,
-      DaviIconValueMapper<DATA>? iconValue,
-      DaviObjectValueMapper<DATA>? objectValue,
+      this.intValue,
+      this.doubleValue,
+      this.stringValue,
+      this.iconValue,
+      this.objectValue,
+      this.summary,
       this.resizable = true,
       this.cellClip = false,
       this.sortable = true,
@@ -47,11 +49,6 @@ class DaviColumn<DATA> extends ChangeNotifier {
       : id = id ?? DaviColumnId(),
         _width = width,
         _grow = grow != null ? math.max(1, grow) : null,
-        stringValueMapper = stringValue,
-        intValueMapper = intValue,
-        iconValueMapper = iconValue,
-        doubleValueMapper = doubleValue,
-        objectValueMapper = objectValue,
         dataComparator = dataComparator ??
             _buildDataComparator(
                 intValue, doubleValue, stringValue, iconValue, objectValue);
@@ -90,11 +87,13 @@ class DaviColumn<DATA> extends ChangeNotifier {
   /// according to value mappings.
   final DaviDataComparator<DATA> dataComparator;
 
-  final DaviIntValueMapper<DATA>? intValueMapper;
-  final DaviDoubleValueMapper<DATA>? doubleValueMapper;
-  final DaviStringValueMapper<DATA>? stringValueMapper;
-  final DaviObjectValueMapper<DATA>? objectValueMapper;
-  final DaviIconValueMapper<DATA>? iconValueMapper;
+  final DaviIntValueMapper<DATA>? intValue;
+  final DaviDoubleValueMapper<DATA>? doubleValue;
+  final DaviStringValueMapper<DATA>? stringValue;
+  final DaviObjectValueMapper<DATA>? objectValue;
+  final DaviIconValueMapper<DATA>? iconValue;
+
+  final SummaryBuilder? summary;
 
   /// Indicates whether the cell widget should be clipped.
   final bool cellClip;